time values should, in face, be doubles

+++ source/PixelResponse.h
@@ -34,7 +34,7 @@ namespace Qpix
 
     // take the number of samples we're skipping and see if there are any
     // electrons in this window, pass back the time of the added electron, if so
-    bool DriftCurrentElectrons(const int&, float&);
+    bool DriftCurrentElectrons(const int&, double&);
 
 }
 #endif
