+++ source/ROOTFileManager.h
@@ -55,8 +55,8 @@ namespace Qpix {
 
             std::vector< int > pixel_x_;
             std::vector< int > pixel_y_;
-            std::vector< std::vector < float > > pixel_reset_;
-            std::vector< std::vector < float > > pixel_tslr_;
+            std::vector< std::vector < double > > pixel_reset_;
+            std::vector< std::vector < double > > pixel_tslr_;
             // std::vector< std::vector< std::vector < int > > > pixel_reset_truth_track_id_;
             // std::vector< std::vector< std::vector < double > > > pixel_reset_truth_weight_;
             std::vector< std::vector< std::vector < int > > > pixel_reset_truth_track_id_;
